add plus.me to OmniAuth initializer
The documentation wasn't totally clear on this
and it would be worth looking into. It states that
if you add your own scopes that you need to re-add
email, and profile. In the code example (which was
not in the README, it listed plus.me as well.

Once this was added the client for the calendar
worked.