install material design lite sass gem
Started the project off experimenting with 
Bootstrap 4. While it has a lot of potential I
am going to wait until the beta version to play
with it some more.

I used material design in a project in the past 
and thought it was pretty cool so I am going to
play with it some more.